Update artifact naming convention in build.yml
The code changes in the `build.yml` file modify the naming convention for a Linux MATLAB artifact. Specifically, the artifact name is updated from `MEXlibCZI-linux-x64-$(git describe --always)` to `MATLAB-MEXlibCZI-linux-x64-$(git describe --always)`. This change ensures that the artifact name now includes a clearer indication that it is related to MATLAB.